fix shellcheck complaints

+++ bin/digifeeds/upload_to_s3_test.sh
@@ -4,13 +4,10 @@ bats_load_library bats-support
 bats_load_library bats-assert
 bats_load_library bats-file
 
-
 setup() {
-  load $SHELLMOCK_PATH
+  load "$SHELLMOCK_PATH"
   SCRATCH_PATH="/tmp/upload_to_s3"
-  CONFIG_PATH=$SCRATCH_PATH/upload_to_s3.config
   SUBJECT=main
-
 
   mkdir $SCRATCH_PATH
 
@@ -35,45 +32,45 @@ setup() {
   touch $INPUT_DIR/$BARCODE_2/00000001.tif
 
   ## Config that's in main.
-  input_directory="$INPUT_DIR"
-  processed_directory="$PROCESSED_DIR"
-  digifeeds_bucket="digifeeds_bucket"
-  timestamp=$TIMESTAMP
-  send_metrics="false"
-  APP_ENV="test"
+  export input_directory="$INPUT_DIR"
+  export processed_directory="$PROCESSED_DIR"
+  export digifeeds_bucket="digifeeds_bucket"
+  export timestamp=$TIMESTAMP
+  export send_metrics="false"
+  export APP_ENV="test"
 
   load "$BATS_TEST_DIRNAME/upload_to_s3.sh"
 }
 
 teardown() {
-  rm -r $SCRATCH_PATH
+  rm -r "$SCRATCH_PATH"
 }
 
 @test "It Works" {
-  shellmock new rclone 
+  shellmock new rclone
   shellmock config rclone 0 1:copy regex-3:^digifeeds_bucket:
-  shellmock config rclone 0 1:check regex-2:$INPUT_DIR regex-3:^digifeeds_bucket:
+  shellmock config rclone 0 1:check regex-2:"$INPUT_DIR" regex-3:^digifeeds_bucket:
   run $SUBJECT
-  
+
   assert_success
 
-  assert_file_exists $PROCESSED_DIR/${TIMESTAMP}_${BARCODE_1}.zip
-  assert_file_exists $PROCESSED_DIR/${TIMESTAMP}_${BARCODE_2}.zip
+  assert_file_exists "$PROCESSED_DIR"/"${TIMESTAMP}"_"${BARCODE_1}".zip
+  assert_file_exists "$PROCESSED_DIR"/"${TIMESTAMP}"_"${BARCODE_2}".zip
 
-  assert_dir_exists $PROCESSED_DIR/${TIMESTAMP}_${BARCODE_1}
-  assert_dir_exists $PROCESSED_DIR/${TIMESTAMP}_${BARCODE_2}
+  assert_dir_exists "$PROCESSED_DIR"/"${TIMESTAMP}"_"${BARCODE_1}"
+  assert_dir_exists "$PROCESSED_DIR"/"${TIMESTAMP}"_"${BARCODE_2}"
   shellmock assert expectations rclone
 }
 
 @test "It filters the appropriate files" {
-  shellmock new rclone 
+  shellmock new rclone
   shellmock config rclone 0 1:copy regex-3:^digifeeds_bucket:
-  shellmock config rclone 0 1:check regex-2:$INPUT_DIR regex-3:^digifeeds_bucket:
+  shellmock config rclone 0 1:check regex-2:"$INPUT_DIR" regex-3:^digifeeds_bucket:
 
   run $SUBJECT
-  cd $BATS_TEST_TMPDIR
-  mv $PROCESSED_DIR/${TIMESTAMP}_${BARCODE_1}.zip ./
-  unzip -q  ${TIMESTAMP}_${BARCODE_1}.zip 
+  cd "$BATS_TEST_TMPDIR"
+  mv "$PROCESSED_DIR/${TIMESTAMP}_${BARCODE_1}.zip" ./
+  unzip -q "${TIMESTAMP}_${BARCODE_1}.zip"
   assert_file_exists '00000001.tif'
   assert_file_exists '00000002.jp2'
   assert_file_exists 'checksum.md5'
@@ -85,15 +82,15 @@ teardown() {
 
 # This test shows that `shopt -s  nullglob` in necessary`
 @test "Emtpy input directory works" {
-  rm -r $INPUT_DIR/$BARCODE_1
-  rm -r $INPUT_DIR/$BARCODE_2
+  rm -r "${INPUT_DIR:?}/${BARCODE_1:?}"
+  rm -r "${INPUT_DIR:?}/${BARCODE_2:?}"
 
   run $SUBJECT
   assert_success
 }
 
 @test "verify_image_order sucess" {
-  run verify_image_order 00000001.tif 00000003.jp2 00000002.tif 
+  run verify_image_order 00000001.tif 00000003.jp2 00000002.tif
   assert_success
 }
 
@@ -103,10 +100,10 @@ teardown() {
 }
 
 @test "Failed image order" {
-  shellmock new rclone 
+  shellmock new rclone
   shellmock config rclone 0 1:copy regex-3:^digifeeds_bucket:
-  shellmock config rclone 0 1:check regex-2:$INPUT_DIR regex-3:^digifeeds_bucket:
-  touch $INPUT_DIR/$BARCODE_1/00000004.jp2
+  shellmock config rclone 0 1:check regex-2:"$INPUT_DIR" regex-3:^digifeeds_bucket:
+  touch "$INPUT_DIR"/"$BARCODE_1"/00000004.jp2
   run $SUBJECT
   assert_output --partial "ERROR: Image order incorrect for $BARCODE_1"
   assert_output --partial "INFO: Total files processed: 1"
@@ -117,8 +114,8 @@ teardown() {
 }
 
 @test "Failed zip" {
-  shellmock new zip 
-  shellmock config zip 1 
+  shellmock new zip
+  shellmock config zip 1
   run $SUBJECT
   assert_output --partial "ERROR: Failed to zip $BARCODE_1"
   assert_output --partial "ERROR: Failed to zip $BARCODE_2"
@@ -130,8 +127,8 @@ teardown() {
 }
 
 @test "Failed copy records error and moves on" {
-  shellmock new rclone 
-  shellmock config rclone 1 1:copy regex-3:^digifeeds_bucket: <<< "Rclone error mock: Failed to copy"
+  shellmock new rclone
+  shellmock config rclone 1 1:copy regex-3:^digifeeds_bucket: <<<"Rclone error mock: Failed to copy"
   run $SUBJECT
   assert_output --partial "ERROR: Failed to copy $BARCODE_1"
   assert_output --partial "ERROR: Failed to copy $BARCODE_2"
@@ -143,9 +140,9 @@ teardown() {
 }
 
 @test "Failed on S3 verification and moves on" {
-  shellmock new rclone 
+  shellmock new rclone
   shellmock config rclone 0 1:copy regex-3:^digifeeds_bucket:
-  shellmock config rclone 1 1:check regex-2:$INPUT_DIR regex-3:^digifeeds_bucket:
+  shellmock config rclone 1 1:check regex-2:"$INPUT_DIR" regex-3:^digifeeds_bucket:
   run $SUBJECT
   assert_output --partial "ERROR: $BARCODE_1 not found in S3"
   assert_output --partial "ERROR: $BARCODE_2 not found in S3"
@@ -157,25 +154,25 @@ teardown() {
 }
 @test "print_metrics" {
   shellmock new pushgateway_advanced
-  shellmock config pushgateway_advanced 0 <<< 5
+  shellmock config pushgateway_advanced 0 <<<5
   run print_metrics 1 2 3 4
   assert_output --partial "aim_digifeeds_upload_to_aws_files_processed_total 6"
   assert_output --partial "aim_digifeeds_upload_to_aws_image_order_errors_total 7"
   assert_output --partial "aim_digifeeds_upload_to_aws_upload_errors_total 8"
   assert_output --partial "aim_digifeeds_upload_to_aws_errors_total 9"
   shellmock assert expectations pushgateway_advanced
-  
+
 }
 
 @test "verify_zip success" {
-   zip_it $INPUT_DIR/$BARCODE_1
-   run verify_zip $INPUT_DIR/$BARCODE_1
-   assert_success
+  zip_it "$INPUT_DIR"/"$BARCODE_1"
+  run verify_zip "$INPUT_DIR"/"$BARCODE_1"
+  assert_success
 }
 
 @test "verify_zip fail" {
-   zip_it $INPUT_DIR/$BARCODE_2
-   mv $INPUT_DIR/$BARCODE_2.zip $INPUT_DIR/$BARCODE_1.zip
-   run verify_zip $INPUT_DIR/$BARCODE_1
-   assert_failure
+  zip_it "$INPUT_DIR"/"$BARCODE_2"
+  mv "$INPUT_DIR"/"$BARCODE_2".zip "$INPUT_DIR"/"$BARCODE_1".zip
+  run verify_zip "$INPUT_DIR"/"$BARCODE_1"
+  assert_failure
 }
